Put after_save back
For some reason `saved_changes` is empty when uploading a hero_image. Seems to have appeared basically from nowhere?
We'll try to fix the issue by putting after_save back - last commit message wasn't super helpful as to why it was
changed in the first place. Possibly coincided with the switch to solid queue?